backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
tensorflow==2.15.0
Pillow==10.2.0
numpy==1.26.4
python-dotenv==1.0.0
